+++ CHANGELOG.md
@@ -5,7 +5,17 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.0 - Unreleased
+## Version 0.2.1 - Unreleased
+
+### Added
+* Slurmify helper script
+* Better slurm support
+
+### Fixed
+* fix `SlurmQueue.is_available` with slurm version 19.x
+
+
+## Version 0.2.0 - Released 2024-06-27
 
 ### Added
 * Add "gpus" as a CLI option